requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
redis==4.3.4
python-socketio==5.4.0
python-dotenv==0.19.0
httpx==0.23.0
google-generativeai==0.3.0
pytest==6.2.5
pytest-asyncio==0.18.3
fakeredis==2.10.0 